完成除商品库存æ­外模块的重构

+++ app/Controller/PermissionController.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Exception\ServiceException;
 use App\Model\Permission;
 use App\Request\PermissionRequest;
 
@@ -24,17 +25,25 @@ class PermissionController extends BaseController
     public function update(PermissionRequest $request)
     {
         $data = $request->validated();
-        $permission = Permission::query()->where('id', $data['id'])->first();
+        $permission = Permission::query()->where('id', $request->route('id'))->first();
+        if (!$permission)
+        {
+            throw new ServiceException(403, '权限不存在');
+        }
         $permission->fill($data);
         $permission->save();
-        return $this->response->json(responseSuccess(200,'更新成功'));
+        return $this->response->json(responseSuccess(200, '更新成功'));
     }
 
     public function delete(PermissionRequest $request)
     {
-        $id = $request->input('id');
+        $id = $request->route('id');
         $permission = Permission::query()->where('id', $id)->first();
+        if (!$permission)
+        {
+            throw new ServiceException(403, '权限不存在');
+        }
         $permission->delete();
-        return $this->response->json(responseSuccess(200,'删除成功'));
+        return $this->response->json(responseSuccess(200, '删除成功'));
     }
 }
