+++ app/Controller/UserAddressesController.php
@@ -40,9 +40,9 @@ class UserAddressesController extends BaseController
         $user = $this->request->getAttribute('user');
 
         $data = $request->validated();
-        $userAddresses = UserAddress::getFirstById($data['id']);
+        $userAddresses = UserAddress::getFirstById($this->request->route('id'));
 
-        if ($user->id != $userAddresses->user_id)
+        if (!$userAddresses || $user->id != $userAddresses->user_id)
         {
             throw new ServiceException(403, '更新地址错误');
         }
@@ -58,11 +58,11 @@ class UserAddressesController extends BaseController
         $user = $this->request->getAttribute('user');
 
         $data = $request->validated();
-        $userAddresses = UserAddress::getFirstById($data['id']);
+        $userAddresses = UserAddress::getFirstById($this->request->route('id'));
 
-        if ($user->id != $userAddresses->user_id)
+        if (!$userAddresses || $user->id != $userAddresses->user_id)
         {
-            throw new ServiceException(403, '更新地址错误');
+            throw new ServiceException(403, '删除地址错误');
         }
 
         $userAddresses->delete();
